backend/sentiment_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
pydantic==2.7.1
python-dotenv==1.0.1
aiokafka>=0.11.0
